+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.4.0] - 2021-05-01
+
 ### Fixed
 
 - Restore target fans speeds after a configuration change.
@@ -59,6 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial version.
 
+[v0.4.0]: https://github.com/musikid/fancy/compare/v0.3.1..v0.4.0
+[v0.3.1]: https://github.com/musikid/fancy/compare/v0.3.0..v0.3.1
 [v0.3.0]: https://github.com/musikid/fancy/compare/fancy-service-0.2.0..v0.3.0
 [v0.2.0]: https://github.com/musikid/fancy/compare/fancy-service-0.1.0...fancy-service-0.2.0
 [v0.1.0]: https://github.com/musikid/fancy/compare/fancy-service-0.1.0